refactor: prepare for move of pkg/resources to machinery
Start enforcing importvet rules, the very first cleanup is breaking
dependency of `pkg/resources` on `pkg/kernel` (only machinery imports
are allowed).

Follow-up PRs will address other problematic imports.
