+++ Makefile
@@ -20,7 +20,7 @@ GOFUMPT_VERSION ?= v0.1.0
 STRINGER_VERSION ?= v0.1.5
 DEEPCOPY_GEN_VERSION ?= v0.21.3
 VTPROTOBUF_VERSION ?= 81d623a9a700ede8ef765e5ab08b3aa1f5b4d5a8
-IMPORTVET ?= autonomy/importvet:f6b07d9
+IMPORTVET ?= ghcr.io/talos-systems/importvet:c9424fe
 OPERATING_SYSTEM := $(shell uname -s | tr "[:upper:]" "[:lower:]")
 TALOSCTL_DEFAULT_TARGET := talosctl-$(OPERATING_SYSTEM)
 INTEGRATION_TEST_DEFAULT_TARGET := integration-test-$(OPERATING_SYSTEM)
