+++ internal/app/machined/pkg/controllers/runtime/kernel_param_defaults_test.go
@@ -14,7 +14,7 @@ import (
 
 	runtimecontrollers "github.com/talos-systems/talos/internal/app/machined/pkg/controllers/runtime"
 	"github.com/talos-systems/talos/internal/app/machined/pkg/runtime"
-	"github.com/talos-systems/talos/pkg/kernel"
+	"github.com/talos-systems/talos/pkg/machinery/kernel"
 	runtimeresource "github.com/talos-systems/talos/pkg/resources/runtime"
 )
