+++ pkg/resources/.importvet.yaml
@@ -1,0 +1,23 @@
+---
+# temporary rules to facilitate moving `pkg/resources` into `pkg/machinery`:
+#  - no imports of anything from Talos except for machinery and pkg/resources itself
+#  - (not enforced yet) external dependencies we don't have to see in the machinery
+rules:
+  - regexp: ^github.com/talos-systems/talos
+    action: deny
+  - regexp: ^github.com/talos-systems/talos/pkg/resources
+    action: allow
+  - regexp: ^github.com/talos-systems/talos/pkg/machinery
+    action: allow
+  #- regexp: ^k8s.io/
+  #  action: deny
+  #- regexp: ^github.com/jxskiss/base62
+  #  action: deny
+  #- regexp: ^github.com/mdlayher/netlink
+  #  action: deny
+  #- regexp: ^github.com/mdlayher/netx
+  #  action: deny
+  #- regexp: ^github.com/prometheus/procfs
+  #  action: deny
+  #- regexp: ^golang.zx2c4.com/wireguard/wgctrl
+  #  action: deny
